+++ package.json
@@ -29,7 +29,7 @@
     "prettier:fix": "yarn prettier --write",
     "format": "yarn prettier:fix && yarn lint:fix",
     "preview": "vite preview",
-    "test:unit": "vitest"
+    "test": "vitest"
   },
   "dependencies": {
     "styled-components": "^5.3.11"
@@ -51,6 +51,7 @@
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-styled-components-a11y": "^2.1.32",
+    "jsdom": "^24.1.0",
     "prettier": "2.8.8",
     "prettier-eslint": "^15.0.1",
     "react": "^18.2.0",
