+++ src/components/UserMessageWithBodyInput.tsx
@@ -7,7 +7,7 @@ import Avatar from '@uikit/ui/Avatar';
 import Label, { LabelColors, LabelTypography } from '@uikit/ui/Label';
 
 import { SentTime } from './MessageComponent';
-import { formatCreatedAtToAMPM } from '../utils';
+import { formatCreatedAtToAMPM } from '../utils/messageTimestamp';
 
 const Root = styled.div`
   display: flex;
