+++ src/utils/index.ts
@@ -5,21 +5,6 @@ import { parseMessageDataSafely } from './messages';
 import { Source } from '../components/SourceContainer';
 import { widgetServiceName } from '../const';
 
-/**
- * createdAt is in UTC. Convert to local time using toLocaleString.
- */
-export function formatCreatedAtToAMPM(createdAt: number) {
-  const date: Date = new Date(createdAt);
-
-  const options: Intl.DateTimeFormatOptions = {
-    hour: 'numeric',
-    minute: 'numeric',
-    hour12: true, // ensures the time is formatted in 12-hour clock notation with AM/PM.
-  };
-  const localTimeString: string = date.toLocaleTimeString(undefined, options);
-  return localTimeString;
-}
-
 // Fixme: Add more languages
 export enum Languages {
   typescript = 'typescript',
